Move daemon startup orchestration out of types.rs
